Replace color selector by elgato classic style preset.
