+++ StreamAwesome/src/components/settings/IconSettings.vue
@@ -3,28 +3,11 @@ import type { CustomIcon } from '@/model/customIcon'
 import PresetOptions from '@/components/settings/PresetOptions.vue'
 import GeneralOptions from '@/components/settings/GeneralOptions.vue'
 import DownloadButton from '@/components/settings/DownloadButton.vue'
-import type { FontAwesomeFamily, FontAwesomeStyle } from '@/model/fontAwesomeConstants'
-import { reactive } from 'vue'
-const props = defineProps({
+defineProps({
   icon: {
     type: Object as () => CustomIcon
   }
 })
-
-const currentIcon = reactive(props.icon ?? ({} as CustomIcon))
-
-// TODO: More refactoring required
-function updateSize(size: number) {
-  currentIcon.fontSize = size
-}
-
-function updateFamily(family: FontAwesomeFamily) {
-  currentIcon.fontAwesomeIcon.family = family
-}
-
-function updateStyle(style: FontAwesomeStyle) {
-  currentIcon.fontAwesomeIcon.style = style
-}
 
 defineEmits<{
   downloadIcon: []
@@ -34,12 +17,7 @@ defineEmits<{
 <template>
   <PresetOptions :icon="icon" />
 
-  <GeneralOptions
-    :icon="icon"
-    @updateSize="updateSize"
-    @updateFamily="updateFamily"
-    @updateStyle="updateStyle"
-  />
+  <GeneralOptions :icon="icon" />
 
   <DownloadButton @downloadIcon="$emit('downloadIcon')" />
 </template>
